Provides error for GDPR topics during registration
Fixes issue 363 - for the GDPR topics, there's no registration via API,
hence there's no corresponding value for the associated GDPR topics.

register first checks to see if there's an existing registration but
fails to confirm that the response is valid and just assumes there's a
'data' key in the response.  This trips up the registration for the GDPR
topics as well as any other invalid topic, e.g., "NONSENSE_TOPIC".